+++ pkg/cmd/run.go
@@ -28,7 +28,6 @@ import (
 	"github.com/tetratelabs/getenvoy/pkg/binary/envoy"
 	"github.com/tetratelabs/getenvoy/pkg/binary/envoy/debug"
 	"github.com/tetratelabs/getenvoy/pkg/globals"
-	ioutil "github.com/tetratelabs/getenvoy/pkg/util/io"
 )
 
 // NewRunCmd create a command responsible for starting an Envoy process
@@ -95,11 +94,8 @@ func InitializeRunOpts(o *globals.GlobalOpts, reference string) error {
 // This is exposed for re-use in "getenvoy extension run"
 func Run(o *globals.GlobalOpts, cmd *cobra.Command, args []string) error {
 	r := envoy.NewRuntime(&o.RunOpts)
-	r.IO = ioutil.StdStreams{
-		In:  cmd.InOrStdin(),
-		Out: cmd.OutOrStdout(),
-		Err: cmd.ErrOrStderr(),
-	}
+	r.Out = cmd.OutOrStdout()
+	r.Err = cmd.ErrOrStderr()
 	debug.EnableAll(r)
 	return r.Run(cmd.Context(), args)
 }
